+++ CHANGELOG.rst
@@ -2,6 +2,8 @@ Changelog
 ---------
 
 v6.4 (??? 2025)
+   - Draw canvas as last element so that UI controls are displayed as
+     long as possible.
    - Get size from fullscreen window to deal with multiple monitors.
 
 v6.3 (Nov 2025)